move partition_id.h
